# SMR-Oriented Optimizations