--- a/include/openmc/particle.h
+++ b/include/openmc/particle.h
@@ -134,6 +134,35 @@ struct NuclideMicroXS {
                             //!< * temperature (eV))
 };
 
+struct NuclideMicroXSNoDepletion {
+  // Microscopic cross sections in barns
+  double total;            //!< total cross section
+  double absorption;       //!< absorption (disappearance)
+  double fission;          //!< fission
+  double nu_fission;       //!< neutron production from fission
+
+  double elastic;          //!< If sab_frac is not 1 or 0, then this value is
+                           //!<   averaged over bound and non-bound nuclei
+  double thermal;          //!< Bound thermal elastic & inelastic scattering
+  double thermal_elastic;  //!< Bound thermal elastic scattering
+  double photon_prod;      //!< microscopic photon production xs
+
+  // Indicies and factors needed to compute cross sections from the data tables
+  int index_grid;        //!< Index on nuclide energy grid
+  int index_temp;        //!< Temperature index for nuclide
+  double interp_factor;  //!< Interpolation factor on nuc. energy grid
+  int index_sab {-1};    //!< Index in sab_tables
+  int index_temp_sab;    //!< Temperature index for sab_tables
+  double sab_frac;       //!< Fraction of atoms affected by S(a,b)
+  bool use_ptable;       //!< In URR range with probability tables?
+
+  // Energy and temperature last used to evaluate these cross sections.  If
+  // these values have changed, then the cross sections must be re-evaluated.
+  double last_E {0.0};      //!< Last evaluated energy
+  double last_sqrtkT {0.0}; //!< Last temperature in sqrt(Boltzmann constant
+                            //!< * temperature (eV))
+};
+
 class NuclideMicroXSCache {
   public:
   NuclideMicroXS neutron_xs_[NEUTRON_XS_SIZE]; //!< Microscopic neutron cross sections