--- a/include/openmc/tallies/tally.h
+++ b/include/openmc/tallies/tally.h
@@ -135,10 +135,6 @@ public:
 
   int deriv_ {C_NONE}; //!< Index of a TallyDerivative object for diff tallies.
 
-private:
-  //----------------------------------------------------------------------------
-  // Private data.
-
   vector<int32_t> filters_; //!< Filter indices in global filters array
 
   //! Index strides assigned to each filter to support 1D indexing.