--- a/src/device_alloc.cpp
+++ b/src/device_alloc.cpp
@@ -217,6 +217,46 @@ void move_read_only_data_to_device()
   " # Fissionable Materials with >= 200 Nuclides: " << n_over_200 << std::endl <<
   " # Fissionable Materials with  < 200 Nuclides: " << n_under_200 << std::endl;
 
+
+  // Determine size of inner dimension for serialized material vectors
+  for (int i = 0; i < model::materials_size; i++) {
+    auto& mat = model::materials[i];
+    model::materials_nuclide.stretch(mat.nuclide_);
+    model::materials_element.stretch(mat.element_);
+    model::materials_atom_density.stretch(mat.atom_density_);
+    model::materials_p0.stretch(mat.p0_);
+    model::materials_mat_nuclide_index.stretch(mat.mat_nuclide_index_);
+    model::materials_thermal_tables.stretch(mat.thermal_tables_);
+  }
+  
+  // Allocate serialized material vectors
+  model::materials_nuclide.resize2d(model::materials_size);
+  model::materials_element.resize2d(model::materials_size);
+  model::materials_atom_density.resize2d(model::materials_size);
+  model::materials_p0.resize2d(model::materials_size);
+  model::materials_mat_nuclide_index.resize2d(model::materials_size);
+  model::materials_thermal_tables.resize2d(model::materials_size);
+  
+  // Populate serialized material vectors
+  for (int i = 0; i < model::materials_size; i++) {
+    auto& mat = model::materials[i];
+    model::materials_nuclide.copy_row(i, mat.nuclide_);
+    model::materials_element.copy_row(i, mat.element_);
+    model::materials_atom_density.copy_row(i, mat.atom_density_);
+    model::materials_p0.copy_row(i, mat.p0_);
+    model::materials_mat_nuclide_index.copy_row(i, mat.mat_nuclide_index_);
+    model::materials_thermal_tables.copy_row(i, mat.thermal_tables_);
+  }
+
+  // Map serialized material vectors to device
+  model::materials_nuclide.copy_to_device();
+  model::materials_element.copy_to_device();
+  model::materials_atom_density.copy_to_device();
+  model::materials_p0.copy_to_device();
+  model::materials_mat_nuclide_index.copy_to_device();
+  model::materials_thermal_tables.copy_to_device();
+
+  /*
   // Prepare serial materials
   int max_nuclides = 0;
   int max_elements = 0;
@@ -284,6 +324,7 @@ void move_read_only_data_to_device()
   if (model::serial_materials_thermal_tables_size > 0) {
     #pragma omp target enter data map(to: model::serial_materials_thermal_tables[:model::serial_materials_thermal_tables_size])
   }
+  */
 
   // Source Bank ///////////////////////////////////////////////////////