## Introduction 

This PR introduces a variety of optimizations and cleanups that resulted from performance analysis of the SMR challenge problem. These optimizations are useful for a wide variety of problems, but have particular impact for the SMR due to that problem having lots of material regions and a fairly large tally space.

## Specialization of neutron XS lookup functions

This optimizations splits the neutron XS functions into multiple versions with/without depletion XS lookups. This does have a noticeable impact on performance (about 10% on the A100). While previously we had simply not been performing the depletion lookups if not needed in the XS kernel, we still had to have a number of variables passed around between functions which causes additional memory usage/movement, as well as additional register usage in the kernel. I did some experimentation with this and did find that it was optimal to actually break the functions up into separate versions so that the types of data passed back could be reduced for inactive batches or active batches that don't require depletion cross section data. Currently, this means that a lot of code is duplicated! Ideally we could find a way to do the different types of lookups efficiently without this high-level specialization, but for now it seems this is an easy enough fix.

One nice side-effect of this optimization is we can view the kernel timings from utilities like `iprof` and `nvprof` in inactive vs. active iterations of the code as they are launched at a high level as two different kernels.

## Changes to XS lookup queue sorting policy

We now sort non-fuel XS lookups in addition to fuel XS lookups. Notably, fuel materials are not sorted by material type (as depleted problems may have O(100k+) material types). However, non-fuel materials are sorted by material type primary, with the energy sort being secondary. This has a nice speedup in the range of 5-10% for A100 where the on-device CUDA sorting is very fast, so the additional overhead of the sort is outweighed by the speedup of non-fuel material lookups. This speedup is most apparent in the SMR which has many more non-fuel XS lookup events as compared to the HM large model we had traditionally been using.

## Serialization of material data

As our working SMR model has about 200k material types, and each material type originally involved mapping of up to 7 pointers, the model in aggregate was mapping over 1.4 million pointers to device with OpenMP. This took several minutes of mapping on all devices. While having this many pointers did not have a kernel performance on some GPUs, on others it created odd performance issues with some key OpenMP operations. To both speed up mapping and avoid these performance issues, it seemed like a good idea to rethink how we were mapping our `Material` objects to device.

To this end, I have added a new serialized `openmc::vector2d` class that inherits from our `openmc::vector` class. The new class adds a few utility functions with the purpose of converting many vector objects into a single 2D vector, where each original vector is copied into a different row of the 2D vector.

To represent six of the vector fields in the `Material` class, we now have six different top level global `vector2d` objects:

```C++
vector2d<int> materials_nuclide;
vector2d<int> materials_element;
vector2d<double> materials_atom_density;
vector2d<int> materials_p0;
vector2d<int> materials_mat_nuclide_index;
vector2d<ThermalTable> materials_thermal_tables;
```

Material data gets initialized normally and stored in the vector fields of the `Material` class. However, after program init when mapping data to device, we do some analysis on the materials array and copy relevant data in from the relevant fields in the `Material` object:

```C++
  //----------------------------------------------------------------------------
  // Data
  ...
  vector<int> nuclide_; //!< Indices in nuclides vector
  vector<int> element_; //!< Indices in elements vector
  xt::xtensor<double, 1> atom_density_; //!< Nuclide atom density in [atom/b-cm]
  ...
  vector<int> p0_; //!< Indicate which nuclides are to be treated with iso-in-lab scattering
 ...
  vector<int> mat_nuclide_index_;
  ...
  vector<ThermalTable> thermal_tables_;
  ...
```

and store it in the new `vector2d` objects for serialization.

So, to be clear, the original fields in the `Material` object still exist, we just do not map them to device. Rather, we copy their data into rows of the top level `vector2d` objects, which occurs in `device_alloc.cpp`

Now, when a particle needs to access this data (e.g., in the neutron XS lookup routine), instead of reading from the `Material` vector fields, it instead will read from the global `vector2d` object. To make refactoring as easy as possible, we define accessor functions in the `Material` class to do this for us, such that throughout OpenMC we only need to refactor from:

```C++
int i_nuclide = mat.nuclide_[i];
double atom_density = mat.device_atom_density_[i];
```

we now just do:

```C++
int i_nuclide = mat.nuclide(i); // uses accessors that will read from the global vector2d for this array
double atom_density = mat.atom_density(i);
```

At the end of the day, the use of this method means we only have to map a few pointers rather than millions, and the logic for converting the vector fields to global vector2d could be followed pretty easily for other classes if we really wanted to.

However, I will note that there are a few shortcomings with this method:

- We still retain the original vector fields in the `Material` class (though they don't get mapped to device). Fully eliminating them is of course possible, but that would require perhaps more involved refactoring and pain for the program file i/o and initialization routines.
- The other problem with my method here is that there is nothing stopping people from accessing the vestigial `Material` vectors in device code which would be bad.
- There is a decent number of lines of new code in the `device_alloc.cpp` file to get everything serialized, but it is at least very straightforward code as the logic for doing the key operations is contained in the `vector2d` class.

Anyway, let me know if someone sees a cleaner way of doing this, I'm not at all deadset on the strategy!

## Performance

Performance improvements on an A100 GPU (in terms of kparticles/sec) for this PR:

  | Old | This PR | Speedup
-- | -- | -- | --
HM Large -  1 MPI Rank - Inactive | 364 | 409 | 1.12
HM Large -  4 MPI Rank - Inactive | 397 | 449 | 1.13
SMR - 1 MPI Rank - Active | 165 | 171 | 1.04
SMR - 4 MPI Rank - Active | 161 | 182 | 1.13

Note the SMR performance numbers are for active batches with multipole XS and 200k unique material regions, whereas the HM Large problem is for inactive batches with pointwise XS and a single fuel material.