@jtramm 2022-10-07
A few additional small changes I forgot to mention:
- Increasing to C++17 from C++14 which is required for the newer version of OneAPI DPL (to support on-device sorting with Intel GPUs)
- Addition of event counters. OpenMC now reports the total number of kernel calls of each type along with the total number of particle events of each type. This was useful in determining that the SMR particle histories are longer than the HM histories. Note that it only reports the data from rank 0 -- it is not reducing the results or taking an average across ranks or anything.
- Cleanup of MPI output so that only rank 0 reports device mapping info etc.