@paulromano 2022-10-26
@jtramm Sorry for the huge delay here. I'm finally past most of the urgent issues on my plate and am starting to work through the backlog of issues and reviews including this. One immediate question -- how difficult would it be to split this up into several PRs (ideally one for each optimization)?